Backend refactor leave sending user ids

+++ Server/MessageAppServer/Controllers/MessageController.cs
@@ -1,11 +1,11 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Server.Services.Chat.MessageService;
+﻿using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Server.Model.Chat;
 using Server.Model.Requests.Message;
 using Server.Model.Responses.Message;
+using Server.Services.Chat.MessageService;
 using Server.Services.Chat.RoomService;
-using Server.Services.User;
 
 namespace Server.Controllers;
 
@@ -13,8 +13,7 @@ namespace Server.Controllers;
 public class MessageController(
     IMessageService messageService,
     IRoomService roomService,
-    ILogger<MessageController> logger,
-    IUserServices userServices
+    ILogger<MessageController> logger
     ) : ControllerBase
 {
     [HttpGet("GetMessages/{roomId}"), Authorize(Roles = "User, Admin")]
@@ -48,19 +47,16 @@ public class MessageController(
             {
                 return BadRequest(ModelState);
             }
+            
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
             var roomIdToGuid = new Guid(request.RoomId);
             if (!roomService.ExistingRoom(roomIdToGuid).Result)
             {
                 return NotFound($"There is no room with this id: {request.RoomId}");
             }
-
-            if (!userServices.ExistingUser(request.UserId).Result)
-            {
-                return NotFound($"There is no user with this id: {request.UserId}");
-            }
         
-            var result = await messageService.SendMessage(request);
+            var result = await messageService.SendMessage(request, userId!);
         
             return Ok(result);
         }
@@ -72,7 +68,7 @@ public class MessageController(
     }
     
     [HttpPatch("EditMessage"), Authorize(Roles = "User, Admin")]
-    public async Task<ActionResult<MessageResponse>> ModifyMessage([FromBody]EditMessageRequest request)
+    public async Task<ActionResult<MessageResponse>> EditMessage([FromBody]EditMessageRequest request)
     {
         try
         {
@@ -80,10 +76,18 @@ public class MessageController(
             {
                 return BadRequest(ModelState);
             }
+            
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userGuid = new Guid(userId!);
 
-            if (!messageService.MessageExisting(request.Id).Result)
+            if (!await messageService.MessageExisting(request.Id))
             {
                 return NotFound($"There is no message with this given id: {request.Id}");
+            }
+            
+            if (!await messageService.UserIsTheSender(userGuid, request.Id))
+            {
+                return BadRequest("You are not supposed to edit other user's messages.");
             }
             
             var result = await messageService.EditMessage(request);
@@ -106,14 +110,11 @@ public class MessageController(
             {
                 return BadRequest(ModelState);
             }
-
-            var messageIdToGuid = new Guid(request.MessageId);
-            if (!messageService.MessageExisting(messageIdToGuid).Result)
-            {
-                return NotFound($"There is no message with this given id: {request.MessageId}");
-            }
-
-            var result = await messageService.EditMessageSeen(request);
+            
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userGuid = new Guid(userId!);
+            
+            var result = await messageService.EditMessageSeen(request, userGuid);
 
             return Ok(result);
         }
@@ -129,10 +130,17 @@ public class MessageController(
     {
         try
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userGuid = new Guid(userId!);
             var idToGuid = new Guid(id);
             if (!messageService.MessageExisting(idToGuid).Result)
             {
                 return NotFound($"There is no message with this given id: {id}");
+            }
+
+            if (!await messageService.UserIsTheSender(userGuid, idToGuid))
+            {
+                return BadRequest("You are not allowed to delete other users messages.");
             }
             
             await messageService.DeleteMessage(idToGuid);
